+++ CHANGES.md
@@ -1,6 +1,14 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+
+### Changed
+
+* Change compression algorithm features flags. [#2250]
+
+[#2250]: https://github.com/actix/actix-web/pull/2250
+
+## 4.0.0-beta.7 - 2021-06-17
 ### Added
 * `HttpServer::worker_max_blocking_threads` for setting block thread pool. [#2200]